Project panel: show embedded takes as 3D clips; SkeletalAnimator; native playback time.
Strip bone rows from expanded models; virtual ::subanim paths use the animclip3d inspector.
Rename Animator3D to SkeletalAnimator (keep Animator3D alias). Push runtime clip time to SkinnedMeshRenderer for future skinning.

Made-with: Cursor
