+++ python/Infernux/core/animation_clip3d.py
@@ -38,6 +38,9 @@ class AnimationClip3D:
     speed: float = 1.0
     loop: bool = True
 
+    # Optional seconds (authoring or tooling); 0.0 = unknown. Embedded takes may be unknown.
+    duration_hint: float = 0.0
+
     file_path: str = field(default="", repr=False, compare=False)
 
     # ── Serialization ───────────────────────────────────────────────
@@ -52,6 +55,7 @@ class AnimationClip3D:
             "bind_pose_bone_names": list(self.bind_pose_bone_names),
             "speed": float(self.speed),
             "loop": bool(self.loop),
+            "duration_hint": float(self.duration_hint),
         }
 
     @classmethod
@@ -68,6 +72,7 @@ class AnimationClip3D:
             bind_pose_bone_names=[str(x) for x in bones],
             speed=float(d.get("speed", 1.0)),
             loop=bool(d.get("loop", True)),
+            duration_hint=float(d.get("duration_hint", 0.0) or 0.0),
         )
 
     def copy(self) -> "AnimationClip3D":
@@ -77,6 +82,10 @@ class AnimationClip3D:
         if not isinstance(other, AnimationClip3D):
             return NotImplemented
         return self.to_dict() == other.to_dict()
+
+    @property
+    def is_valid_reference(self) -> bool:
+        return bool((self.source_model_guid or "").strip() or (self.source_model_path or "").strip())
 
     # ── File I/O ─────────────────────────────────────────────────────
 
@@ -108,8 +117,67 @@ class AnimationClip3D:
         except (OSError, json.JSONDecodeError, KeyError, TypeError, ValueError):
             return None
 
-    # ── Helpers ───────────────────────────────────────────────────────
+    @classmethod
+    def from_embedded_take_virtual_path(cls, virtual_path: str) -> Optional["AnimationClip3D"]:
+        """Build a read-only clip view for ``model.fbx::subanim:<index>`` (Project Panel)."""
+        token = "::subanim:"
+        if token not in virtual_path:
+            return None
+        base, _, rest = virtual_path.partition(token)
+        if not base.strip():
+            return None
+        try:
+            idx = int(rest.strip())
+        except ValueError:
+            return None
+        # Placeholder / overflow row from the project panel
+        if idx < 0 or idx >= 999999:
+            return None
+        if not os.path.isfile(base):
+            return None
 
-    @property
-    def is_valid_reference(self) -> bool:
-        return bool((self.source_model_guid or "").strip() or (self.source_model_path or "").strip())
+        from Infernux.core.asset_types import read_meta_file
+
+        meta = read_meta_file(base) or {}
+        csv = (meta.get("animation_names_csv") or "")
+        if isinstance(csv, str):
+            names = [p.strip() for p in csv.split(",") if p.strip()]
+        else:
+            names = []
+
+        if idx >= len(names):
+            return None
+
+        take_name = names[idx]
+        guid = _read_asset_guid_from_meta_sidecar(base)
+        bind_csv = (meta.get("bone_names_csv") or "")
+        if isinstance(bind_csv, str):
+            bind_names = [p.strip() for p in bind_csv.split(",") if p.strip()]
+        else:
+            bind_names = []
+
+        clip = cls(
+            name=take_name,
+            source_model_guid=guid,
+            source_model_path=os.path.normpath(base),
+            take_name=take_name,
+            bind_pose_bone_names=bind_names,
+            speed=1.0,
+            loop=True,
+            duration_hint=0.0,
+        )
+        clip.file_path = virtual_path
+        return clip
+
+
+def _read_asset_guid_from_meta_sidecar(asset_path: str) -> str:
+    """Return GUID from a ``.meta`` file's root (not the metadata{} map)."""
+    meta_path = asset_path + ".meta"
+    if not os.path.isfile(meta_path):
+        return ""
+    try:
+        with open(meta_path, "r", encoding="utf-8") as f:
+            root = json.load(f)
+        return str(root.get("guid", "") or "")
+    except (OSError, json.JSONDecodeError, TypeError, ValueError):
+        return ""
